cheri: Fix compile in drivers/block/
- Standard uintptr_t vs. unsigned long issues.
- Standard ioctl issues.
- Need to call cheri_make_kernel_data_cap() in drivers/drbd/.
  This use of peer provided data look dubious in the first place
  so put a FIXCHERI in there for now.

+++ drivers/block/drbd/drbd_nl.c
@@ -3280,7 +3280,7 @@ put_result:
 	err = resource_statistics_to_skb(skb, &resource_statistics, !capable(CAP_SYS_ADMIN));
 	if (err)
 		goto out;
-	cb->args[0] = (long)resource;
+	cb->args[0] = (intptr_t)resource;
 	genlmsg_end(skb, dh);
 	err = 0;
